requirements.txt
langchain
langchain-community
langchain-openai
langchain-experimental
streamlit
pandas
matplotlib
seaborn
pygwalker
jupyterlab
tiktoken
tabulate
python-dotenv
langchain-google-genai
mysql-connector-python